+++ client/src/components/forms/index.js
@@ -1,0 +1,2 @@
+export { default as CategoryForm } from './CategoryForm'
+export {default as LocalSearch} from './LocalSearch'
